Rename remaining camelCase files to kabab-case
