+++ examples/context/components/__fixtures__/Color/black-and-white.js
