+++ examples/context/components/__fixtures__/Color/blue-and-yellow.json
