+++ examples/local-state/components/Counter/__fixtures__/five-clicks.js
