+++ examples/normalize-props-proxy/components/Counter/__fixtures__/five-clicks.js
