+++ packages/react-cosmos-component-playground/src/components/FixtureList/__fixtures__/nested-data.js
