+++ packages/react-cosmos-component-playground/src/components/FixtureList/__tests__/data-mapper.js
@@ -1,4 +1,4 @@
-import fixturesToTreeData from '../dataMapper';
+import fixturesToTreeData from '../data-mapper';
 
 const input = {
   'dirA/Component1': ['fixtureA', 'fixtureB'],
