+++ packages/react-cosmos-component-playground/src/components/FixtureList/data-mapper.js
