+++ packages/react-cosmos-component-playground/src/components/SvgIcon/__fixtures__/down-arrow.js
