+++ packages/react-cosmos-component-playground/src/components/SvgIcon/__fixtures__/right-arrow.js
